auto duration handling
apparently the "default auto archive duration" is a hint given to clients, which are then supposed to pass it on when creating a thread. the api has its own default, which is always a day, and if your client doesn't override that it's what gets used. So now we fetch channel info once in a while.